+++ python/jaeger/actor/commands/wago.py
@@ -6,14 +6,24 @@
 # @Filename: wago.py
 # @License: BSD 3-clause (http://www.opensource.org/licenses/BSD-3-Clause)
 
+import asyncio
+
 import click
 
+from clu.parser import pass_args
 from . import jaeger_parser
 
 
 @jaeger_parser.group()
-def wago():
-    pass
+@pass_args()
+def wago(command, fps):
+
+    wago = fps.wago
+    if wago is None or not wago.connected:
+        command.failed(text='WAGO not connected.')
+        raise click.Abort()
+
+    return
 
 
 @wago.command()
@@ -21,9 +31,6 @@ async def status(command, fps):
     """Outputs the status of the PLCs."""
 
     wago = fps.wago
-    if wago is None:
-        command.failed(text='WAGO not set up.')
-        return
 
     for category in wago.list_categories():
         measured = await wago.read_category(category)
@@ -37,15 +44,15 @@ async def status(command, fps):
 @click.option('--on/--off', default=None,
               help='the value of the PLC. If not provided, '
                    'switches the current status.')
-async def switch(command, fps, plc, on):
+@click.option('--cycle', is_flag=True, help='power cycles a relay. '
+                                            'The final status is on.')
+async def switch(command, fps, plc, on, cycle):
     """Switches the status of an on/off PLC."""
 
-    desired_status = None
-
     wago = fps.wago
-    if wago is None:
-        command.failed(text='WAGO not set up.')
-        return
+
+    if cycle:
+        on = False
 
     try:
         plc_obj = wago.get_plc(plc)
@@ -67,20 +74,25 @@ async def switch(command, fps, plc, on):
                                 'is not on or off.')
             return
 
-    if on is True:
-        await wago.turn_on(plc_obj.name)
-        desired_status = 'on'
-    elif on is False:
-        await wago.turn_off(plc_obj.name)
-        desired_status = 'off'
-
-    # Do a sanity check
-    status = await plc_obj.read(convert=True)
-
-    if not status == desired_status:
+    try:
+        if on is True:
+            await wago.turn_on(plc_obj.name)
+        elif on is False:
+            await wago.turn_off(plc_obj.name)
+    except Exception:
         command.failed(text=f'failed to set status of PLC {plc_obj.name}.')
         return
 
-    command.done(text=f'PLC {plc_obj.name!r} is now {desired_status!r}.')
+    if cycle:
+        command.write('d', text='waiting 1 second before powering up.')
+        await asyncio.sleep(1)
+        try:
+            await wago.turn_on(plc_obj.name)
+        except Exception:
+            command.failed(text=f'failed to power PLC {plc_obj.name} back on.')
+            return
+
+    status = await plc_obj.read(convert=True)
+    command.done(text=f'PLC {plc_obj.name!r} is now {status!r}.')
 
     return